Stuck on CORS error trying to join a game
I'm not sure why this is failing. I've tried a few things on the server
to make it accept requests with any headers, but with it as it's
currently set up it works if I edit and resend one of the requests in
dev tools and manually write out the body.

It seems from that that the missing request body is what's causing the
problem, but I'm not sure if that's just omitted when the preflight
options request fails. It's returning a 403.